refactor: copy Empty implementation from cmdk(#34)
