+++ components/ui/multiple-selector.tsx
@@ -3,15 +3,9 @@
 import * as React from 'react';
 import { X } from 'lucide-react';
 
-import {
-  Command,
-  CommandGroup,
-  CommandItem,
-  CommandEmpty,
-  CommandList,
-} from '@/components/ui/command';
-import { Command as CommandPrimitive } from 'cmdk';
-import { useEffect } from 'react';
+import { Command, CommandGroup, CommandItem, CommandList } from '@/components/ui/command';
+import { Command as CommandPrimitive, useCommandState } from 'cmdk';
+import { useEffect, forwardRef } from 'react';
 import { Badge } from '@/components/ui/badge';
 import { cn } from '@/lib/utils';
 
@@ -125,6 +119,33 @@ function removePickedOption(groupOption: GroupOption, picked: Option[]) {
   }
   return cloneOption;
 }
+
+/**
+ * The `CommandEmpty` of shadcn/ui will cause the cmdk empty not rendering correctly.
+ * So we create one and copy the `Empty` implementation from `cmdk`.
+ *
+ * @reference: https://github.com/hsuanyi-chou/shadcn-ui-expansions/issues/34#issuecomment-1949561607
+ **/
+const CommandEmpty = forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<typeof CommandPrimitive.Empty>
+>(({ className, ...props }, forwardedRef) => {
+  const render = useCommandState((state) => state.filtered.count === 0);
+
+  if (!render) return null;
+
+  return (
+    <div
+      ref={forwardedRef}
+      className={cn('py-6 text-center text-sm', className)}
+      cmdk-empty=""
+      role="presentation"
+      {...props}
+    />
+  );
+});
+
+CommandEmpty.displayName = 'CommandEmpty';
 
 const MultipleSelector = React.forwardRef<MultipleSelectorRef, MultipleSelectorProps>(
   (
